// src/app/home/store/home.action.type.ts
export enum HomeActionType {
  GLOBAL_START = '[HOME] GLOBAL_START',
  GLOBAL_SUCCESS = '[HOME] GLOBAL_SUCCES',
  GLOBAL_ERROR = '[HOME] GLOBAL_ERROR',

  YOUR_START = '[HOME] YOUR_START',
  YOUR_SUCCESS = '[HOME] YOUR_SUCCESS',
  YOUR_ERROR = '[HOME] YOUR_ERROR',

  POPULAR_TAGS_START = '[HOME]  POPULAR TAGS_START',
  POPULAR_TAGS_SUCCESS = '[HOME] POPULAR TAGS_SUCCESS',
  POPULAR_TAGS_ERROR = '[HOME] POPULAR TAGS_ERROR',

  TAG_START = '[HOME]  TAG_START',
  TAG_SUCCESS = '[HOME] TAG_SUCCESS',
  TAG_ERROR = '[HOME] TAG_ERROR',
}